Improve error handling with session detection in mixin fetch API
The fetch family of methods on the GraphLayerMixin now handle errors so
that the calling context doesn't have to worry about uncaught promises.
It also now checks to see if a session cookie is available before
issuing a call in order to head off cases where the user is not signed
in. (Only relevant if the component displays content that is non-shared.)