Improve drive component drive name formatting
When loading a drive associated with a group, we do not need to pull
additional group info if the group name can be located via the 'owner'
property on the drive object.

For sites, we still pull additional site info to obtain a better site
name for labeling the drive.