Ensure event list queries all viewable events
We needed to include events whose start date did not occur in the range
but whose end date did occur in the range. Note that to my knowledge it
is not feasible using the list item api to query events that completely
overlap the range (i.e. whose start and end dates occur occur outside
the range).