Implement by-day day of and weekday of recurrences
For the monthlyByDay and yearlyByDay repeat rules, there were a few
extra patterns to implement that were not originally documented in the
XML schema. These include the day and weekday patterns.

The previous assumption was that these rules only worked with
day-of-week attributes, which was causing an infinite loop in the
generation method. This patch fixes the infinite loop and prevents it
from occurring in case any future, unknown patterns crop up.