Add all day event support to list events calendar functionality
For an all-day event, we parse the date string differently. This fixes
an issue where the Microsoft platform does not return the start date for
an all-day event with midnight adjusted for the site timezone, which
resulted in an all-day event having an incorrect start time after the
timezone conversion.

This commit also improves the formatting of the CalendarEvent component
to account for all-day events. Date formats were also improved to be
more grammatically correct.